Joystick: Map joystick circle to large square.
This commit uses an elliptical unit grid mapping for the joystick, which allows the circular stick inputs to be mapped to a square shape. This provides better use of the small joysticks on a gamepad controller.

Caveat: This will cause issues if pitch+roll and yaw+throttle are not mapped on the same sticks. An option may be added to enable joystick square mapping.